feat: ProjectListItemButtonsWebsiteAndShare for mobile moved below the description
