+++ client/src/components/Project/ProjectListItem/ProjectListItem.module.scss
@@ -59,3 +59,9 @@
     }
   }
 }
+
+.buttonsWebsiteAndShare {
+  @media #{$desktop-up} {
+    display: none;
+  }
+}
