+++ client/src/components/Project/ProjectListItemButtonsWebsiteAndShare/types.ts
@@ -1,0 +1,9 @@
+export default interface ProjectListItemButtonsWebsiteAndShareProps {
+  address: string;
+  className?: string;
+  name?: string;
+  website?: {
+    label?: string;
+    url?: string;
+  };
+}
