+++ client/src/components/Project/ProjectListItemHeader/ProjectListItemHeader.tsx
@@ -1,21 +1,16 @@
-import cx from 'classnames';
-import React, { FC, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import React, { FC } from 'react';
 import { useParams } from 'react-router-dom';
 
+import ProjectListItemButtonsWebsiteAndShare from 'components/Project/ProjectListItemButtonsWebsiteAndShare';
 import ButtonAddToAllocate from 'components/shared/ButtonAddToAllocate';
 import Img from 'components/ui/Img';
-import Svg from 'components/ui/Svg';
-import Tooltip from 'components/ui/Tooltip';
 import env from 'env';
 import useIdsInAllocation from 'hooks/helpers/useIdsInAllocation';
 import useIsAllocatedTo from 'hooks/helpers/useIsAllocatedTo';
 import useCurrentEpoch from 'hooks/queries/useCurrentEpoch';
 import useIsDecisionWindowOpen from 'hooks/queries/useIsDecisionWindowOpen';
 import useUserAllocations from 'hooks/queries/useUserAllocations';
-import { ROOT_ROUTES } from 'routes/RootRoutes/routes';
 import useAllocationsStore from 'store/allocations/store';
-import { arrowTopRight, share } from 'svg/misc';
 
 import styles from './ProjectListItemHeader.module.scss';
 import ProjectListItemHeaderProps from './types';
@@ -28,8 +23,6 @@ const ProjectListItemHeader: FC<ProjectListItemHeaderProps> = ({
   epoch,
 }) => {
   const { ipfsGateways } = env;
-  const { t, i18n } = useTranslation('translation', { keyPrefix: 'views.project' });
-  const [isLinkCopied, setIsLinkCopied] = useState(false);
   const { epoch: epochUrl } = useParams();
   const { data: currentEpoch } = useCurrentEpoch();
   const { data: userAllocations } = useUserAllocations(epoch);
@@ -59,31 +52,6 @@ const ProjectListItemHeader: FC<ProjectListItemHeaderProps> = ({
     userAllocations,
   );
 
-  const onShareClick = (): boolean | Promise<boolean> => {
-    const { origin } = window.location;
-    const url = `${origin}${ROOT_ROUTES.project.absolute}/${epochUrl}/${address}`;
-
-    if ((window.navigator.share as any) && !window.navigator.userAgent.includes('Macintosh')) {
-      window.navigator.share({
-        title: i18n.t('meta.fundrasingOnOctant', {
-          projectName: name,
-        }),
-        url,
-      });
-
-      return false;
-    }
-
-    return window.navigator.clipboard.writeText(url).then(() => {
-      setIsLinkCopied(true);
-
-      setTimeout(() => {
-        setIsLinkCopied(false);
-      }, 1000);
-      return true;
-    });
-  };
-
   return (
     <div className={styles.root}>
       <Img
@@ -95,33 +63,7 @@ const ProjectListItemHeader: FC<ProjectListItemHeaderProps> = ({
         {name}
       </div>
       <div className={styles.wrapper}>
-        <div className={styles.actionButtonsWrapper}>
-          <a
-            className={cx(styles.element, styles.actionButton, styles.websiteLink)}
-            href={website?.url}
-            rel="noreferrer"
-            target="_blank"
-          >
-            <span className={styles.labelOrUrl}>{website!.label || website!.url}</span>
-            <Svg classNameSvg={styles.arrowTopRightIcon} img={arrowTopRight} size={1} />
-          </a>
-          <Tooltip
-            className={styles.element}
-            hideAfterClick
-            onClickCallback={onShareClick}
-            text={isLinkCopied ? i18n.t('common.copied') : i18n.t('common.copy')}
-            variant="small"
-          >
-            <button className={styles.actionButton} type="button">
-              <span>{t('share')}</span>
-              <Svg
-                classNameSvg={cx(styles.shareIcon, isLinkCopied && styles.isCopied)}
-                img={share}
-                size={3.2}
-              />
-            </button>
-          </Tooltip>
-        </div>
+        <ProjectListItemButtonsWebsiteAndShare address={address} name={name} website={website} />
         <ButtonAddToAllocate
           className={styles.buttonAddToAllocate}
           dataTest="ProjectListItemHeader__ButtonAddToAllocate"
